@@ -1,6 +1,6 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-	  <TargetFrameworks>netstandard2.1;net6.0;net8.0</TargetFrameworks>
+	  <TargetFrameworks>netstandard2.1;net8.0</TargetFrameworks>
 	  <LangVersion>12.0</LangVersion>
 	  <AssemblyName>DarkPeakLabs.RdapClient</AssemblyName>
 
