@@ -31,7 +31,7 @@ public abstract class RdapBootstrap
     /// </summary>
     /// <param name="urls">list of urls</param>
     /// <returns></returns>
-    protected static Uri SelectUrl(IReadOnlyCollection<Uri> urls)
+    protected static Uri SelectUrl(IReadOnlyList<Uri> urls)
     {
         _ = urls ?? throw new ArgumentNullException(paramName: nameof(urls));
 
@@ -42,7 +42,7 @@ public abstract class RdapBootstrap
                 return url;
             }
         }
-        return urls.First();
+        return urls[0];
     }
 
     protected async Task<T> GetOrAddCacheItemAsync<T>(string key, Func<Task<T>> addItemFactory)
