@@ -11,15 +11,15 @@ public class RdapObjectTagsServiceRegistryItem
     /// <summary>
     /// Contact information for the registered service provider identifiers
     /// </summary>
-    public IReadOnlyCollection<string> Contacts { get; set; }
+    public IReadOnlyList<string> Contacts { get; set; }
 
     /// <summary>
     /// Alphanumeric identifiers that identify RDAP service providers
     /// </summary>
-    public IReadOnlyCollection<string> Identifiers { get; set; }
+    public IReadOnlyList<string> Identifiers { get; set; }
 
     /// <summary>
     /// Base RDAP URLs usable for the entries
     /// </summary>
-    public IReadOnlyCollection<Uri> ServiceUrls { get; set; }
+    public IReadOnlyList<Uri> ServiceUrls { get; set; }
 }
