@@ -1,11 +1,11 @@
-﻿using DarkPeakLabs.Rdap.Values.Json;
+﻿using DarkPeakLabs.Rdap.Values;
 using System.Collections.Generic;
 
 namespace DarkPeakLabs.Rdap
 {
     public class RdapContactPhoneNumber
     {
-        public IReadOnlyCollection<RdapPhoneNumberType> Types { get; set; }
+        public IReadOnlyList<RdapPhoneNumberType> Types { get; set; }
         public string Value { get; set; }
     }
 }
