@@ -1,4 +1,4 @@
-﻿using DarkPeakLabs.Rdap.Values.DnsSec;
+﻿using DarkPeakLabs.Rdap.Values;
 using System.Text.Json.Serialization;
 
 namespace DarkPeakLabs.Rdap
@@ -20,7 +20,7 @@ namespace DarkPeakLabs.Rdap
         /// an integer as specified by the algorithm field of a DNS DS record as described by RFC 4034 in presentation format
         /// </summary>
         [JsonPropertyName("algorithm")]
-        public AlgorithmType? Algorithm { get; set; }
+        public DnsSecAlgorithmType? Algorithm { get; set; }
 
         /// <summary>
         /// a string as specified by the digest field of a DNS DS record as specified by RFC 4034 in presentation format
@@ -32,6 +32,6 @@ namespace DarkPeakLabs.Rdap
         /// an integer as specified by the digest type field of a DNS DS record as specified by RFC 4034 in presentation format
         /// </summary>
         [JsonPropertyName("digestType")]
-        public DigestType? DigestType { get; set; }
+        public DnsSecDigestType? DigestType { get; set; }
     }
 }
