@@ -1,4 +1,4 @@
-﻿using DarkPeakLabs.Rdap.Values.DnsSec;
+﻿using DarkPeakLabs.Rdap.Values;
 using System.Text.Json.Serialization;
 
 namespace DarkPeakLabs.Rdap
@@ -14,7 +14,7 @@ namespace DarkPeakLabs.Rdap
         /// an integer representing the flags field value in the DNSKEY record RFC4034 in presentation format
         /// </summary>
         [JsonPropertyName("flags")]
-        public DnsKeyFlags? Flags { get; set; }
+        public DnsSecKeyFlags? Flags { get; set; }
 
         /// <summary>
         /// an integer representation of the protocol field value of the DNSKEY record RFC4034 in presentation format
@@ -32,6 +32,6 @@ namespace DarkPeakLabs.Rdap
         /// an integer as specified by the algorithm field of a DNSKEY record as described by RFC 4034 in presentation format
         /// </summary>
         [JsonPropertyName("algorithm")]
-        public AlgorithmType? Algorithm { get; set; }
+        public DnsSecAlgorithmType? Algorithm { get; set; }
     }
 }
