@@ -1,4 +1,4 @@
-﻿using DarkPeakLabs.Rdap.Values.Json;
+﻿using DarkPeakLabs.Rdap.Values;
 using System.Collections.Generic;
 using System.Text.Json.Serialization;
 
@@ -13,7 +13,7 @@ namespace DarkPeakLabs.Rdap
         /// an array of strings, with each string denoting the relationship between the variants and the containing domain object
         /// </summary>
         [JsonPropertyName("relation")]
-        public IReadOnlyCollection<RdapDomainVariantRelation> Relation { get; set; }
+        public IReadOnlyList<RdapDomainVariantRelation> Relation { get; set; }
 
         /// <summary>
         /// the name of the Internationalized Domain Name (IDN) table of codepoints, such as one listed with the IANA (see http://www.iana.org/domains/idn-tables)
@@ -25,6 +25,6 @@ namespace DarkPeakLabs.Rdap
         /// an array of variant name objects
         /// </summary>
         [JsonPropertyName("variantNames")]
-        public IReadOnlyCollection<RdapDomainVariantName> VariantNames { get; set; }
+        public IReadOnlyList<RdapDomainVariantName> VariantNames { get; set; }
     }
 }
