@@ -1,58 +1,38 @@
-﻿using DarkPeakLabs.Rdap.Conformance;
-using Microsoft.Extensions.Logging;
-using System;
-using System.Globalization;
+﻿using System.Globalization;
 using System.Text.Json;
-using System.Text.Json.Serialization;
+using System.Text.Json.Nodes;
 
 namespace DarkPeakLabs.Rdap.Serialization
 {
-    internal class RdapStringConverter : JsonConverter<string>
+    internal static class RdapStringConverter
     {
-        /// <summary>
-        /// Class logger
-        /// </summary>
-        private ILogger _logger;
-
-        /// <summary>
-        /// RDAP conformance
-        /// </summary>
-        private readonly RdapConformance conformance;
-
-        /// <summary>
-        /// constructor
-        /// </summary>
-        /// <param name="errors"></param>
-        public RdapStringConverter(RdapConformance conformance, ILogger logger = null)
+        public static bool TryGetValue(JsonValue jsonValue, RdapSerializerContext context, out string value)
         {
-            this.conformance = conformance;
-            _logger = logger;
-        }
+            value = default;
 
-        public override string Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-        {
-            switch (reader.TokenType)
+            var valueKind = jsonValue.GetValueKind();
+            switch (valueKind)
             {
-                case JsonTokenType.String:
-                    return reader.GetString();
+                case JsonValueKind.String:
+                    value = jsonValue.GetValue<string>();
+                    return true;
 
-                case JsonTokenType.Number:
-                    conformance.AddJsonViolation(RdapConformanceViolationSeverity.Warning, ref reader, "Expected data type of JSON value is string, found a number instead.");
-                    return reader.GetInt32().ToString(CultureInfo.InvariantCulture);
-
-                case JsonTokenType.StartArray:
-                    conformance.AddJsonViolation(RdapConformanceViolationSeverity.Warning, ref reader, "Expected data type of JSON value is string, found an array instead.");
-                    reader.Skip();
-                    return null;
+                case JsonValueKind.Number:
+                    context.AddJsonViolationError(jsonValue, $"Property {jsonValue.GetPropertyName()} should be a string not a number");
+                    if (jsonValue.TryGetValue<int>(out var intValue))
+                    {
+                        value = intValue.ToString(CultureInfo.InvariantCulture);
+                        return true;
+                    }
+                    else
+                    {
+                        return false;
+                    }
 
                 default:
-                    throw new RdapJsonException($"Unexpected token type {reader.TokenType} when reading string", ref reader);
+                    context.AddJsonViolationError(jsonValue, $"{valueKind} is not a valid JSON token type for string property {jsonValue.GetPropertyName()}");
+                    return false;
             }
-        }
-
-        public override void Write(Utf8JsonWriter writer, string value, JsonSerializerOptions options)
-        {
-            throw new NotImplementedException();
         }
     }
 }
