@@ -1,12 +1,12 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace DarkPeakLabs.Rdap.Values.DnsSec
+namespace DarkPeakLabs.Rdap.Values
 {
     [Flags]
 #pragma warning disable CA1711 // Identifiers should not have incorrect suffix
 #pragma warning disable CA2217 // Do not mark enums with FlagsAttribute
-    public enum DnsKeyFlags : int
+    public enum DnsSecKeyFlags
 #pragma warning restore CA2217 // Do not mark enums with FlagsAttribute
 #pragma warning restore CA1711 // Identifiers should not have incorrect suffix
     {
