@@ -1,5 +1,4 @@
-﻿using System.Text.Json.Nodes;
-using DarkPeakLabs.Rdap.Serialization;
+﻿using DarkPeakLabs.Rdap.Serialization;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace DarkPeakLabs.Rdap.Test
@@ -11,18 +10,14 @@ namespace DarkPeakLabs.Rdap.Test
         public void Test1()
         {
             RdapSecureDns rdapSecureDns = RdapSerializer.Deserialize<RdapSecureDns>(ReadJsonFile("rdap-secure-dns-1.json"));
+            Assert.IsNull(rdapSecureDns.DnsDsRecords);
         }
 
         [TestMethod]
         public void Test2()
         {
             RdapSecureDns rdapSecureDns = RdapSerializer.Deserialize<RdapSecureDns>(ReadJsonFile("rdap-secure-dns-2.json"));
-        }
-
-        [TestMethod]
-        public void Test3()
-        {
-            JsonNode node = RdapSerializer.Deserialize<JsonNode>(ReadJsonFile("rdap-secure-dns-2.json"));
+            Assert.IsNotNull(rdapSecureDns.DnsDsRecords);
         }
     }
 }
